requirements.txt
google-generativeai
sqlite3
gradio
python-dotenv
jsonschema